+++ client/client.go
@@ -48,6 +48,9 @@ func DefaultClients(nc *nats.Conn) (*RunGroup, error) {
 	metrics := NewManager(nc, NewMetricsClient, nil)
 	g.Add(metrics)
 
+	modbus := NewManager(nc, NewModbusClient, nil)
+	g.Add(modbus)
+
 	particle := NewManager(nc, NewParticleClient, nil)
 	g.Add(particle)